Auto-generate respec HTML with GitHub Actions (#479)
Automatically generate the respec HTML on every commit, using respec2html to automate the process. Then push the results to the gh-pages branch.

This cuts down on the effort required to make edits to the spec, and avoids issues where we may forget to update index.html or one of the other numerous generated HTML files.

The generated files are now no longer stored in the master branch of the repo, and will only appear in the gh-pages branch.