[Editorial] Drop "div" around IDL sections
The use of `<div>` around sections that define IDL seems to confuse ReSpec as
it takes out the headings included in that section from the table of contents.

If that's intended, the right way to exclude headings from the table of
contents would be to flag the headings with a `notoc` class:
https://respec.org/docs/#notoc-class

The `<div>` seem useless in any case. This update gets rid of them.