Enable auto-publication
This automates the publication of all 9 EME documents to the `gh-pages` branch
and to /TR, using the w3c/spec-prod action.

The `W3CTRMANIFEST` files are no longer needed as a result and nothing needs to
be Git-ignored anymore either.

Also note the change of shortname for Encrypted Media Extensions to
`encrypted-media-2`.