Auto-create auto-publish workflows
Automate all the things!

This adds an auto-publish workflow template and a companion script to generate
auto-publish workflows for each and every spec in the repository that only run
when needed, meaning only when the spec source did change. The script only
needs to run once in a while when new specs get added or when the workflows
need to change.

Script and template are in the `.github` folder to avoid polluting the repo
with build tools.