Enable auto-publication (#553)
This automates the publication of all 9 EME documents to the `gh-pages` branch
and to /TR, using the w3c/spec-prod action.

In particular, this adds an auto-publish workflow template and a companion script to generate auto-publish workflows for each and every spec in the repository that only run when needed, meaning only when the spec source did change. The script only needs to run once in a while when new specs get added or when the workflows need to change.

Script and template are in the `.github` folder to avoid polluting the repo with build tools.

The `W3CTRMANIFEST` files are no longer needed as a result and nothing needs to be Git-ignored anymore either.

Also note the change of shortname for Encrypted Media Extensions to `encrypted-media-2`.