Note loosening impl requirements is out of scope in SOTD
Concerns were raised during the review of the new Media Working Group charter
about the possible impact of the features being added to EME on user privacy,
interoperability, royalty-free implementation of the standard and access to
media.

This adjusts the Status of this Document section to clarify that loosening of
implementation requirements that were published in the W3C Recommendation is
per charter out of scope for the Media Working Group.